glx: allow querying GLX_WIDTH/HEIGHT of a Window
The GLX spec allows vanilla Window or Pixmap XIDs to be treated as
GLXDrawables, although it also states that this is only for backwards
compatibility with GLX 1.2. It's ambiguous whether apis added since GLX
1.2 must also accept a Window or Pixmap as a GLXDrawable.

Since the proprietary Nvidia driver does in fact allow applications to
pass a Window XID to glXQueryDrawable to at least query the GLX_WIDTH
and GLX_HEIGHT attributes and since this has caused numerous
compatibility problems with projects that have only been tested with
Nvidia's driver, this patch now allows passing a Window or Pixmap XID
when making a GLXGetDrawableAttributes request which will only report
the width and height as if calling XGetGeometry.

https://bugs.freedesktop.org/show_bug.cgi?id=54080

Signed-off-by: Robert Bragg <robert@sixbynine.org>