glx: Allow passing a Window to GLXGetDrawableAttributes
This updates __glXGetDrawable so that it does not require a context
pointer to be able to create a __GLXdrawable when a vanilla Window XID
is used as a GLXDrawable.

Removing this restriction means that DoGetDrawableAttributes can also
use this helper function allowing clients to pass Window XIDs as
GLXDrawables to glXQueryDrawable().

https://bugs.freedesktop.org/show_bug.cgi?id=54080
https://forums.oculus.com/viewtopic.php?f=20&t=16201
https://forums.oculus.com/viewtopic.php?f=34&t=16664

Signed-off-by: Robert Bragg <robert@sixbynine.org>